fix: [BE] 방정리 smembers to scan
